Update ConfigServiceProvider.php
Could not merge a config key that accepted both a string or array value (default provided was a string, my value was an array). This is my proposed solution. Cannot say with confidence it is elegant or desirable to allow this however. Please close PR if you do not agree or see a valid use case for this behavior.

Also, some CS/PSR2 thingies "fixed".